added variable stun for roll
stun time is now based on time left during roll. so if you roll right infront of the wall, you get the full 0.5 seconds of stun, but if you only hit the wall near the end of your roll, you get only 0.1 seconds. this makes it less punishing and more strategic