changed artwork, tweaked controller
changed to the sample artwork for level and player. because it is bigger, it means that we had to change up all of the values to make them fit again. it isn't perfect, but its okay. started working on the collision detection, can now detect when fireworks / goal is made, but not both. will probably remove goal as part of tileset.